apps/backend-python/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
PyYAML==6.0.2
pydantic==2.10.6
pydantic-settings==2.8.0
aiosqlite==0.20.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
websockets==13.1
